for some reason the iscsi and userspace files did not get updated from my scsi protocol checkin so here they are again

+++ usr/stgtd.h
@@ -24,7 +24,7 @@ extern int nl_cmnd_call(int fd, int type, char *data, int size, int *res);
 extern int ipc_open(void);
 extern void ipc_event_handle(int fd);
 
-extern int scsi_cmnd_process(int tid, uint32_t lun, uint8_t *scb,
+extern int scsi_cmnd_process(int tid, uint64_t lun, uint8_t *scb,
 			     uint8_t *data, int *len);
 
 #endif
